+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Updated
 
 - Add support for InputfieldPage (Experimental)
+- Build Mystique fields inside hook by using ProcessPageEdit::buildForm() hook method
 
 ### Removed
 
@@ -10,6 +11,7 @@
 
 ### Added
 
+- InputfieldMystique.js added for javascript processes
 - InputfieldPage fields returns Page, PageArray, NullPage or false values
 - getDataArray() for MystiqueValue class, return pure data array
 - Import and Export Mystique field values